move eventual asserts to core module
