+++ python-libraries/nanover-core/tests/imd/test_imd_client.py
@@ -1,6 +1,4 @@
 import itertools
-import time
-from typing import Callable
 
 import pytest
 from nanover.imd.particle_interaction import ParticleInteraction
@@ -9,35 +7,12 @@ from nanover.imd.imd_state import (
     interaction_to_dict,
     dict_to_interaction,
 )
+from nanover.testing import assert_in_soon, assert_not_in_soon, assert_equal_soon
 from nanover.utilities.change_buffers import DictionaryChange
 
 from .test_imd_server import imd_server_client, imd_server, interaction
 
 IMMEDIATE_REPLY_WAIT_TIME = 0.01
-
-
-def assert_equal_soon(a: Callable, b: Callable, interval=0.1, timeout=1.0):
-    __tracebackhide__ = True
-    deadline = time.monotonic() + timeout
-    while time.monotonic() < deadline and a() != b():
-        time.sleep(interval)
-    assert a() == b()
-
-
-def assert_in_soon(a: Callable, b: Callable, interval=0.1, timeout=1.0):
-    __tracebackhide__ = True
-    deadline = time.monotonic() + timeout
-    while time.monotonic() < deadline and a() not in b():
-        time.sleep(interval)
-    assert a() in b()
-
-
-def assert_not_in_soon(a: Callable, b: Callable, interval=0.1, timeout=1.0):
-    __tracebackhide__ = True
-    deadline = time.monotonic() + timeout
-    while time.monotonic() < deadline and a() in b():
-        time.sleep(interval)
-    assert a() not in b()
 
 
 def test_start_interaction(imd_server_client):
